% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_determination.R
\name{build.all.femaleness}
\alias{build.all.femaleness}
\title{Give the list of femaleness in a population}
\usage{
build.all.femaleness(genome)
}
\description{
Give the list of femaleness in a population
}
