% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{build.all.genotype}
\alias{build.all.genotype}
\title{Build all genotype for a given genome}
\usage{
build.all.genotype(genome)
}
\description{
a genotype is defined by two haplotype. Since an
haplotype can be defined by its index, i.e. a number,
a genotype can be seen as two number. This function return
all the possible genotype for a given genome, i.e. all the possible
pairs of number. Notice that the pair is not order so is c(1,2) is
equivalent to c(2,1) and only one is return. Moreover, this function
check that only compatible haplotype are combine. Therefore genotype
of the form YY would not be returned.
}
