% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{build.all.haplotype}
\alias{build.all.haplotype}
\title{Build all possible haplotype}
\usage{
build.all.haplotype(genome)
}
\description{
For a given genome, various haplotype are possible.
This function build all the possible haplotype
for a given genome
}
