% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{build.all.possible.gamete}
\alias{build.all.possible.gamete}
\title{Build all possible gamete}
\usage{
build.all.possible.gamete(genome)
}
\description{
For a given genome, various gamete are possible.
This function build all the possible gamete
for a given genome
}
