% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_modifier.R
\name{build.all.recombination.modifier}
\alias{build.all.recombination.modifier}
\title{build the recombination factor of all genotype}
\usage{
build.all.recombination.modifier(genome)
}
\description{
build the recombination factor of all genotype
}
