% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{build.genotype.from.locus}
\alias{build.genotype.from.locus}
\title{For a locus, give all possible genotype}
\usage{
build.genotype.from.locus(genome, locus)
}
\description{
For a genome and a locus of this genome, we expect
to have different possible genotype like aa, aA and AA.
In the simple case, this function just return all the possible
combination of allele. However, in some case, some configuration
are not possible and this function takes care of eliminating them
like yy is not possible.
}
