% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{compute.frequency.evolution}
\alias{compute.frequency.evolution}
\title{Compute the evolution of the frequency of various genotype}
\usage{
compute.frequency.evolution(genome, initial.frequency = NULL,
  generations = 25)
}
\arguments{
\item{genome}{A S4 object of the type genome}

\item{initial.frequency}{The initial frequency of the various genotype. If NULL is given
the initial frequencies will all be set to the same value}

\item{generations}{The number of generation that have to be computed (including the first one)}
}
\value{
A matrix containing the frequencies of each genotype at each generation
}
\description{
Given a genome, this function simulate the evolution of the frequency
of all the possible genotype that exist
}
\examples{
locus1 = data.frame(allele1=c(1,1),allele2 = c(1,2),sd = c(0,1),fitness.male=c(1,1),fitness.female=c(1,1))
locus2 = data.frame(allele1=  c(1,1,2),allele2 = c(1,2,2),fitness.female = c(1,0.9,0.8),fitness.male = c(0.6,0.8,1))
genome = create.genome(locus1,locus2)
freqs <- compute.frequency.evolution(genome)
}
