% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_object.R
\docType{class}
\name{genome-class}
\alias{genome-class}
\alias{create.genome}
\title{Create a genome}
\arguments{
\item{locus}{A list of locus}

\item{male.recombination}{Optional: A vector containing the recombination rate between the different locus. It should be of length n-1 where n is the number of locus.}

\item{female.recombination}{Optional: A vector containing the recombination rate between the different locus. It should be of length n-1 where n is the number of locus.}
}
\description{
A genome is a S4 class which define all the properties of a genome, it is basically
a list of locus
}
\examples{
locus1 = create.locus(chrom1=c(1,1),chrom2 = c(1,2),sd = c(0,1),fitness.male=c(1,1),fitness.female=c(1,1))
locus2 = create.locus(chrom1=  c(1,1,2),chrom2 = c(1,2,2),fitness.female = c(1,0.9,0.8),fitness.male = c(0.6,0.8,1))
genome = create.genome(locus=list(locus1,locus2))
}
