% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{genotype.is.in.locus}
\alias{genotype.is.in.locus}
\title{check if a certain genotype configuration is in a locus}
\usage{
genotype.is.in.locus(allele.locus, locus.all.config)
}
\description{
check if a certain genotype configuration is in a locus
}
