% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_modifier.R
\name{get.all.recombination.modifier}
\alias{get.all.recombination.modifier}
\title{get the recombination factor of all genotype}
\usage{
get.all.recombination.modifier(genome)
}
\description{
get the recombination factor of all genotype
}
