% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get.allele.frequency}
\alias{get.allele.frequency}
\title{Return the allele frequency for a given locus}
\usage{
get.allele.frequency(genome, freqs, locus.position)
}
\arguments{
\item{genome}{A S4 object of type genome}

\item{freqs}{a matrix of frequency as returned by the function `compute.frequency.evolution`}
}
\description{
given a matrix of frequency returned by the function `compute.frequency.evolution`
and the associated genome, return a matrix containing the frequency of each allele
of a given locuc through time. Each row contains an allele while each column
contains a generation.
}
\examples{
locus1 = create.locus(chrom1=c(1,1),chrom2 = c(1,2),sd = c(0,1))
locus2 = create.locus(chrom1=  c(1,1,2),chrom2 = c(1,2,2),fitness.female = c(1,0.9,0.8),fitness.male = c(0.6,0.8,1))
genome = create.genome(locus=list(locus1,locus2))
freqs <- compute.frequency.evolution(genome)
freqs.allele <- get.allele.frequency(genome, freqs)
}
