% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_function.R
\name{get.childs.genotype.from.parent.genotype}
\alias{get.childs.genotype.from.parent.genotype}
\title{Get the genotype of all child from their parents}
\usage{
get.childs.genotype.from.parent.genotype(male.genotype, female.genotype, genome)
}
\description{
This function returns a 2x4 matrix containing all the possible
genotype (on each line) of the child coming from two differents
parents.
}
