% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_function.R
\name{get.childs.index.from.parent.index}
\alias{get.childs.index.from.parent.index}
\title{Get the index of all child from their parents index}
\usage{
get.childs.index.from.parent.index(male.genotype, female.genotype, genome)
}
\description{
This function returns the index of all child, i.e. the position
of this individual in the all.genotype matrix.
}
