% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexual_function.R
\name{get.female}
\alias{get.female}
\title{give the list of female in a population}
\usage{
get.female(genome)
}
\description{
This function return the index of all genotype that are,
at least partially, female
}
