% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_determination.R
\name{get.femaleness}
\alias{get.femaleness}
\title{Determine the femaleness of a genotype}
\usage{
get.femaleness(genotype.index, genome)
}
\description{
Gives the proportion of individuals whith the given genotype
which are female
}
