% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{get.fitness.from.genotype.male}
\alias{get.fitness.from.genotype.male}
\title{get fitness of a male}
\usage{
get.fitness.from.genotype.male(genotype, genome, all.haplotype, all.genotype)
}
\description{
get the fitness of a male from the index value of its genotype
}
