% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get.frequency.from.gamete.frequency}
\alias{get.frequency.from.gamete.frequency}
\title{This functions allows to generate genotypic frequency
by specifying the initial frequency og the gamete in male and female}
\usage{
get.frequency.from.gamete.frequency(genome, male.gamete.frequency,
  female.gamete.frequency)
}
\arguments{
\item{genome}{A S4 object of type genome}

\item{male.gamete.frequency}{a list of the frequency of the gamete for male (should sum up to one)}

\item{female.gamete.frequency}{a list of the frequency of the gamete for female (should sum up to one)}
}
\description{
This function is usefull mainly to start a simulation with controlled
gamete frequency
}
