% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_determination.R
\name{get.gamete.female}
\alias{get.gamete.female}
\title{Determine the list of gamete present in female}
\usage{
get.gamete.female(genome)
}
\description{
Determine the list of gamete present in female
}
