% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{get.gamete.from.allele}
\alias{get.gamete.from.allele}
\title{get gamete index from the alleles present on an gamete}
\usage{
get.gamete.from.allele(genome, alleles)
}
\description{
get gamete index from the alleles present on an gamete
}
