% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{get.gamete.from.index}
\alias{get.gamete.from.index}
\title{get an gamete from it's index}
\usage{
get.gamete.from.index(gamete.index, nb.alleles)
}
\arguments{
\item{gamete.index}{The index of the gamete we want to generate}

\item{nb.alleles}{A vector containing the number of allele for each locus}
}
\description{
All possible gamete can be generated and store in a
matrix. This function return the content of the line "index"
of this matrix.
}
