% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_subfunction.R
\name{get.gamete.names.from.allele.names}
\alias{get.gamete.names.from.allele.names}
\title{return names for the gamete if a name is present on a locus}
\usage{
get.gamete.names.from.allele.names(genome)
}
\description{
return names for the gamete if a name is present on a locus
}
