% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_subfunction.R
\name{get.gamete.names.from.allele.number}
\alias{get.gamete.names.from.allele.number}
\title{return names for the gamete if no name is provided}
\usage{
get.gamete.names.from.allele.number(genome)
}
\description{
return names for the gamete if no name is provided
}
