% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{get.genotype.index.from.gametes.index}
\alias{get.genotype.index.from.gametes.index}
\title{get genotype from two gamete}
\usage{
get.genotype.index.from.gametes.index(gametes, all.genotype)
}
\description{
This function return the index value of a genotype
defined by the value of the index of two gametes
}
