% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{get.genotype.index.from.haplotypes.index}
\alias{get.genotype.index.from.haplotypes.index}
\title{get genotype from two haplotype}
\usage{
get.genotype.index.from.haplotypes.index(haplotypes, all.genotype)
}
\description{
This function return the index value of a genotype
defined by the value of the index of two haplotypes
}
