% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get.genotype.names}
\alias{get.genotype.names}
\title{get the name of all genotype present in the population}
\usage{
get.genotype.names(genome)
}
\arguments{
\item{genome}{A S4 object of type genome}
}
\description{
This function returns a name for all genotype present in the population
This is useful for plotting result but also to know in which order the genotype
are store in genome.  This is useful for example to specify the initial frequency
}
\details{
Notice that if allele.name is specified, the name will contain this name and is
therefore much easier to read that if it's not present, where the name of the allele
is just their number.
}
\examples{
locus1 = create.locus(allele1=c(1,1),allele2 = c(1,2),sd = c(0,1),allele.name = c("x","y"))
locus2 = create.locus(allele1=  c(1,1,2),allele2 = c(1,2,2),fitness.female = c(1,0.9,0.8),fitness.male = c(0.6,0.8,1),allele.name = c("F","M"))
genome = create.genome(locus=list(locus1,locus2))
get.genotype.names(genome)
}
