% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{get.genotype.with.given.gamete}
\alias{get.genotype.with.given.gamete}
\title{get all the genotype which contains a given gamete}
\usage{
get.genotype.with.given.gamete(genome, gamete)
}
\description{
get all the genotype which contains a given gamete
}
