% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{get.genotype.with.given.haplotype}
\alias{get.genotype.with.given.haplotype}
\title{get all the genotype which contains a given haplotype}
\usage{
get.genotype.with.given.haplotype(genome, haplotype)
}
\description{
get all the genotype which contains a given haplotype
}
