% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_determination.R
\name{get.haplotype.female}
\alias{get.haplotype.female}
\title{Determine the list of haplotype present in female}
\usage{
get.haplotype.female(genome)
}
\description{
Determine the list of haplotype present in female
}
