% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{get.haplotype.from.allele}
\alias{get.haplotype.from.allele}
\title{get haplotype index from the alleles present on an haplotype}
\usage{
get.haplotype.from.allele(genome, alleles)
}
\description{
get haplotype index from the alleles present on an haplotype
}
