% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{get.haplotype.from.index}
\alias{get.haplotype.from.index}
\title{get an haplotype from it's index}
\usage{
get.haplotype.from.index(haplotype.index, nb.alleles)
}
\arguments{
\item{haplotype.index}{The index of the haplotype we want to generate}

\item{nb.alleles}{A vector containing the number of allele for each locus}
}
\description{
All possible haplotype can be generated and store in a
matrix. This function return the content of the line "index"
of this matrix.
}
