% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_determination.R
\name{get.haplotype.male}
\alias{get.haplotype.male}
\title{Determine the list of haplotype present in male}
\usage{
get.haplotype.male(genome)
}
\description{
Determine the list of haplotype present in male
}
