% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_subfunction.R
\name{get.haplotype.names.from.allele.number}
\alias{get.haplotype.names.from.allele.number}
\title{return names for the haplotype if no name is provided}
\usage{
get.haplotype.names.from.allele.number(genome)
}
\description{
return names for the haplotype if no name is provided
}
