% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_subfunction.R
\name{get.haplotype.names.with.names}
\alias{get.haplotype.names.with.names}
\title{return names for the haplotype if a name is present on a locus}
\usage{
get.haplotype.names.with.names(genome)
}
\description{
return names for the haplotype if a name is present on a locus
}
