% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexual_function.R
\name{get.id.sd.locus}
\alias{get.id.sd.locus}
\title{give the position of the sd locus}
\usage{
get.id.sd.locus(genome)
}
\description{
This function return the position of the sexual determining
locus, i.e. the locus containing the sex determination, in the genome.
If it is not found, return an error.
If more than one sd is found, return an error
}
