% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexual_function.R
\name{get.male}
\alias{get.male}
\title{give the list of male in a population}
\usage{
get.male(genome)
}
\description{
This function return the index of all genotype that are,
at least partially, male
}
