% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_determination.R
\name{get.maleness}
\alias{get.maleness}
\title{Determine the maleness of a genotype}
\usage{
get.maleness(genotype.index, genome)
}
\description{
Gives the proportion of individuals whith the given genotype
which are male
}
