% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{get.nb.alleles.per.locus}
\alias{get.nb.alleles.per.locus}
\title{get the number of allele on each locus}
\usage{
get.nb.alleles.per.locus(genome)
}
\value{
A table containing the number of allele at each locus
}
\description{
get the number of allele on each locus
}
