% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_function.R
\name{get.nb.gamete}
\alias{get.nb.gamete}
\title{get the number of possible gamete that exist for a given genome}
\usage{
get.nb.gamete(genome)
}
\description{
get the number of possible gamete that exist for a given genome
}
