% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{get.nb.genotype}
\alias{get.nb.genotype}
\title{get the number of possible genotype that exist for a given genome}
\usage{
get.nb.genotype(genome)
}
\description{
The number of possible genotype is of the order of the square of the
number of possible haplotype. If all genotype are possible, it's given
by \eqn{n(n+1)/2}, but if some are not (like YY), it might be more complicated
}
