% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{get.nb.haplotype}
\alias{get.nb.haplotype}
\title{get the number of possible haplotype that exist for a given genome}
\usage{
get.nb.haplotype(genome)
}
\description{
get the number of possible haplotype that exist for a given genome
}
