% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_modifier.R
\name{get.recombination.modifier.from.genotype}
\alias{get.recombination.modifier.from.genotype}
\title{get the recombination modifier of a single genotype}
\usage{
get.recombination.modifier.from.genotype(genotype, genome)
}
\description{
get the recombination modifier of a single genotype
}
