% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexual_function.R
\name{is.female}
\alias{is.female}
\title{Determine if a genotype is male}
\usage{
is.female(genotype.index, genome)
}
\description{
This function return true if the genotype with index
genotype.index might generate a male.
}
