% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_determination.R
\name{is.male}
\alias{is.male}
\title{Determine if a genotype is male}
\usage{
is.male(genotype.index, genome)
}
\description{
This function return true if the genotype with index
genotype.index might generate a male and false otherwise.
}
