% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_function.R
\name{plot.allele.frequency}
\alias{plot.allele.frequency}
\title{Plot the evolution of the frequency of an allele}
\usage{
\method{plot}{allele.frequency}(genome, freqs, locus.position)
}
\arguments{
\item{genome}{A S4 object of the type genome}

\item{freqs}{a matrix of frequency as returned by the function `compute.frequency.evolution`}

\item{locus.position}{the index of the locus from which we want to plot the allele frequency}
}
\description{
given a matrix of frequency returned by the function `compute.frequency.evolution`
and the associated genome, plot the frequency of all possible allele at a given locus.
}
\examples{
locus1 = create.locus(chrom1=c(1,1),chrom2 = c(1,2),sd = c(0,1),fitness.male=c(1,1),fitness.female=c(1,1))
locus2 = create.locus(chrom1=  c(1,1,2),chrom2 = c(1,2,2),fitness.female = c(1,0.9,0.8),fitness.male = c(0.6,0.8,1))
genome = create.genome(locus=list(locus1,locus2))
freqs <- compute.frequency.evolution(genome)
plot.allele.frequency(genome, freqs,1)
plot.allele.frequency(genome, freqs,2)

}
