% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.genotype.frequency}
\alias{plot.genotype.frequency}
\title{Plot the evolution of the frequency of genotype}
\usage{
plot.genotype.frequency(genome, freqs)
}
\arguments{
\item{genome}{A S4 object of the type genome}

\item{freqs}{a matrix of frequency as returned by the function `compute.frequency.evolution`}
}
\description{
given a matrix of frequency returned by the function `compute.frequency.evolution`
and the associated genome, plot the frequency of all possible genotype through time
}
\examples{
locus1 = create.locus(allele1=c(1,1),allele2 = c(1,2),sd = c(0,1))
locus2 = create.locus(allele1=  c(1,1,2),allele2 = c(1,2,2),fitness.female = c(1,0.9,0.8),fitness.male = c(0.6,0.8,1))
genome = create.genome(locus=list(locus1,locus2))
freqs <- compute.frequency.evolution(genome)
plot.genotype.frequency(genome, freqs)
}
