% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_function.R
\name{simulate.frequency}
\alias{simulate.frequency}
\title{Get the frequency of new born as a function of adults}
\usage{
\method{simulate}{frequency}(genome, initial.frequency)
}
\description{
Get the frequency of new born as a function of adults
}
