% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{where.is.locus}
\alias{where.is.locus}
\title{gives the position of a certain configuration}
\usage{
where.is.locus(allele.locus, locus.all.config)
}
\description{
gives the position of a certain configuration
}
