% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_function.R
\name{get.fitness.from.genotype.female}
\alias{get.fitness.from.genotype.female}
\title{get fitness of a female}
\usage{
get.fitness.from.genotype.female(genotype, genome, all.haplotype, all.genotype)
}
\description{
get the fitness of a female from the index value of its genotype
}
