% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_determination.R
\name{is.female}
\alias{is.female}
\title{Determine if a genotype is female}
\usage{
is.female(genotype.index, genome)
}
\description{
This function return true if the genotype with index
genotype.index might generate a female, and false otherwise.
}
